[L2B-5863] Move logger init closer to app root
