+++ packages/backend/src/config/Config.ts
@@ -20,8 +20,6 @@ export interface Config {
   readonly name: string
   readonly isReadonly: boolean
   readonly projects: Project[]
-  readonly logger: LoggerConfig
-  readonly logThrottler: LogThrottlerConfig | false
   readonly clock: ClockConfig
   readonly metricsAuth: MetricsAuthConfig | false
   readonly database: DatabaseConfig
