+++ packages/backend/src/index.ts
@@ -1,9 +1,19 @@
 import Bugsnag from '@bugsnag/js'
-import { getEnv } from '@l2beat/backend-tools'
+import {
+  ElasticSearchTransport,
+  ElasticSearchTransportOptions,
+  LogFormatterEcs,
+  LogFormatterJson,
+  LogFormatterPretty,
+  Logger,
+  LoggerOptions,
+  LoggerTransportOptions,
+  getEnv,
+} from '@l2beat/backend-tools'
 
 import { Application } from './Application'
 import { getConfig } from './config'
-import { initializeErrorReporting } from './tools/ErrorReporter'
+import { initializeErrorReporting, reportError } from './tools/ErrorReporter'
 
 main().catch(() => {
   process.exit(1)
@@ -18,22 +28,25 @@ async function main() {
     environment,
   )
 
+  const logger = createLogger(environment)
+
   try {
     const config = getConfig()
-    const app = new Application(config)
+    const app = new Application(config, logger)
     await app.start()
   } catch (e) {
     console.error(e)
+    logger.error('Failed to start the application', e)
 
     if (isErrorReportingEnabled) {
-      await reportError(e)
+      await reportIndexError(e)
     }
 
     throw e
   }
 }
 
-async function reportError(e: unknown) {
+async function reportIndexError(e: unknown) {
   if (typeof e === 'string') {
     Bugsnag.notify(e, (event) => {
       event.context = 'Backend index.ts'
@@ -51,4 +64,51 @@ async function reportError(e: unknown) {
   // wait 10 seconds for the error to be reported
   console.log('Waiting 10 seconds for the error to be reported')
   await new Promise((resolve) => setTimeout(resolve, 10000))
+}
+
+function createLogger(environment: string): Logger {
+  const isLocal = environment === 'local'
+
+  const loggerTransports: LoggerTransportOptions[] = [
+    {
+      transport: console,
+      formatter: isLocal ? new LogFormatterPretty() : new LogFormatterJson(),
+    },
+  ]
+
+  // Elastic Search logging
+  const esEnabled = getEnv().optionalBoolean('ES_ENABLED') ?? false
+
+  if (esEnabled) {
+    console.log('Elastic Search logging enabled')
+    const options: ElasticSearchTransportOptions = {
+      node: getEnv().string('ES_NODE'),
+      apiKey: getEnv().string('ES_API_KEY'),
+      indexPrefix: getEnv().string('ES_INDEX_PREFIX'),
+      flushInterval: getEnv().optionalInteger('ES_FLUSH_INTERVAL'),
+    }
+
+    loggerTransports.push({
+      transport: new ElasticSearchTransport(options),
+      formatter: new LogFormatterEcs(),
+    })
+  }
+
+  const options: Partial<LoggerOptions> = {
+    logLevel: getEnv().string('LOG_LEVEL', 'INFO') as LoggerOptions['logLevel'],
+    utc: isLocal ? false : true,
+    transports: loggerTransports,
+    reportError,
+  }
+
+  let logger = new Logger(options)
+  if (!isLocal) {
+    logger = logger.withThrottling({
+      callsUntilThrottle: 4,
+      clearIntervalMs: 5000,
+      throttleTimeMs: 20000,
+    })
+  }
+
+  return logger
 }
